tests: Ensure that used structures are not higher than the users' Application::apiVersion
`apiVersion` in this context seems to constrain what the user promises
to use in Vulkan, even if the actual instance or physical device
version is higher; this is confirmed by https://docs.vulkan.org/spec/latest/chapters/initialization.html#_extending_physical_device_core_functionality.

A random struct has been used in this example, but it could be anything.